Re-implements Test.HashType as a top level Enum
