+++ src/main/java/com/suse/salt/netapi/calls/modules/Test.java
@@ -32,33 +32,6 @@ public class Test {
             new TypeToken<Map<String, String>>(){});
 
     /**
-     * The type of hash based on Python hashlib - only always supported types.
-     */
-    public enum HashType {
-
-        SHA1("sha1"),
-        MD5("md5"),
-        SHA256("sha256"),
-        SHA224("sha224"),
-        SHA512("sha512"),
-        SHA384("sha384");
-
-        private String hashType;
-
-        HashType(String hashType) {
-            this.hashType = hashType;
-        }
-
-        /**
-         *
-         * @return the corresponding hash type.
-         */
-        public String getHashType() {
-            return hashType;
-        }
-    }
-
-    /**
      * Availability report of all execution modules
      */
     public static class ModuleReport {
